separated test_selenium into utilities and tests, passed pylint, db test untested
